rename(#14): ApiResult data.model로 이동
