+++ app/src/main/java/com/ssafy/neegongnaegong/data/model/ApiResult.kt
@@ -1,4 +1,4 @@
-package com.ssafy.neegongnaegong.data.model.calendar
+package com.ssafy.neegongnaegong.data.model
 
 sealed class ApiResult<out T> {
     data class Success<T>(val data: T) : ApiResult<T>()
